Refactor Zantara endpoint and add dedicated agent
